pyproject.toml
[project]
name = "docmind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "inngest>=0.5.16",
    "llama-index-core>=0.14.15",
    "llama-index-readers-file>=0.5.6",
    "openai>=2.24.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.0",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "flake8>=7.3.0",
    "isort>=8.0.1",
    "pre-commit>=4.5.1",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]